+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-06-25 20:57:13
-// Commit: style(test): harmonize light and dark theme palette for test
+// Updated at: 2026-06-26 12:02:19
+// Commit: fix(test): fix memory leak by cancelling test subscription
 
 class testComponentState {
-  final String commitRef = '2331';
-  final String lastMessage = 'style(test): harmonize light and dark theme palette for test';
-  final DateTime timestamp = DateTime.parse('2026-06-25T20:57:13+05:00');
+  final String commitRef = '2353';
+  final String lastMessage = 'fix(test): fix memory leak by cancelling test subscription';
+  final DateTime timestamp = DateTime.parse('2026-06-26T12:02:19+05:00');
 }